requirements.txt
psycopg2
crispy-bootstrap5==0.4
django-environ==0.9.0
Django==3.2.4
django-bootstrap5==2.0.1
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-filter==2.4.0
django-session-timeout==0.1.0
djangorestframework==3.12.4
social-auth-app-django==5.0.0
social-auth-core==4.1.0
